+++ Core/Maps/IMap.cs
@@ -32,7 +32,7 @@ namespace Helion.Maps
         Sector? GetNextLowestCeiling(Sector sector);
         Sector? GetNextHighestFloor(Sector sector);
         Sector? GetNextHighestCeiling(Sector sector);
-        byte GetMinLightLevelNeighbor(Sector sector);
-        byte GetMaxLightLevelNeighbor(Sector sector);
+        short GetMinLightLevelNeighbor(Sector sector);
+        short GetMaxLightLevelNeighbor(Sector sector);
     }
 }
